Add row_num, keywords, organization, published and thumbnail
to search results.

Caveat: The thumbnails for keyword search results were added rather
awkwardly, mostly due to my lack of JavaScript and Svelte knowledge,
but also the way item.graphicOverview is a string and not a JSON object,
with _double_ double quotes that I needed to do some tweaking before
I could use JSON.parse() on the string:

    JSON.parse(item.graphicOverview.replaceAll('""', '"'))

Let's revisit this at a later date.