Update to latest API with _en and _fr appended to title etc.
Semantic Search API update: Bo Lu (@bo-lu) enhanced the response
so that English and French title, keywords and description are returned
in the same call, as:

  - title_en, title_fr
  - keywords_en, keywords_fr
  - description_en, description_fr

in order to speed up the site when switching languages.

This front-end demo version, still supporting English only, is now bumped
to v0.3.9 (2025-04-30).  French page to come soon.